feat: atomic design の layouts 層を導入し app.vue から抽出
- app/layouts/default.vue: ヘッダー・ナビ付き共通レイアウト（app.vue から抽出）
- app/layouts/auth.vue: ヘッダーなしの認証ページ専用レイアウト
- app.vue を <NuxtLayout><NuxtPage /> のみに簡略化
- auth/login・auth/user-register に definePageMeta({ layout: 'auth' }) を追加
- Storybook の stories パスに app/layouts/** を追加し layouts stories を作成
