+++ app/layouts/auth.vue
@@ -1,0 +1,5 @@
+<template>
+  <div>
+    <slot />
+  </div>
+</template>
